Increased durability and changed recipe
Increased durability to 4800 - now it takes the armor 80 minutes (4 ingame days) to decay from full to 0 (assuming no damage is taken).

Since it was much cheaper to use 7 steel and 1 uranium to craft new fully loaded armor than to refuel it from 0 to full with 4 uranium and 4 levels, the recipe for chestplate was changed to use iron chestplate 4 steel and 4 uranium.