Update to 1.19.4 and changes in damage handling
Added checkDamageType(DamageSource) method to NucleArmorItem which replaced content of an if statement in handleDamage method as it was getting really long.
Now the armor protects against all damage that do not bypass armor and additionally against fall, freeze, drowning, lightning, hot floor, in wall, and falling blocks, anvils and stalactites damage.