Lead instead of steel in crafting
Changed crafting recipes to use lead and iron armor instead of steel. Idea is iron armor is lined with lead to block the radiation. As lead would be a bottom layer it does not wear of so lead is not used for repairs. The top layer of iron means by default it is an iron armor and not powered behaves like one (except it is heavy because of lead layer) and iron is used to repair parts of this armor (except chestplate which is "refueled" with uranium).